Share single Spring context across SBS3 transaction tests
Extract a shared abstract base test (AbstractTransactionMonitoringIntegrationTest)
that owns @SpringBootTest, the @Import of one @TestConfiguration, and the shared
domain/helper classes. TransactionMonitoringBeanPostProcessorTest and
ProxyingDataSourceBeanPostProcessorTest now extend it, so both resolve to an
identical MergedContextConfiguration and share one cached ApplicationContext
(2 contexts -> 1, verified via spring-test-profiler).

TransactionMonitoringEndpointTest is left untouched and keeps its own context.
